+++ frontend/src/pages/admin/CreateEventPage.tsx
@@ -15,7 +15,7 @@ import {
 import { addDays } from 'date-fns';
 import { toast } from 'react-toastify';
 
-import { Button, Input, Card, PasswordGenerator, LocalizedDateInput } from '../../components/common';
+import { Button, Input, Card, PasswordGenerator, LocalizedDateInput, TimeField } from '../../components/common';
 import { ThemeCustomizerEnhanced, GalleryPreview, WelcomeMessageEditor, FeedbackSettings } from '../../components/admin';
 import { CustomerAccountPicker } from '../../components/admin/CustomerAccountPicker';
 import { useMutation, useQuery } from '@tanstack/react-query';
@@ -613,19 +613,15 @@ export const CreateEventPage: React.FC = () => {
               </label>
               {!formData.is_full_day && (
                 <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                  <Input
-                    type="time"
-                    step={900}
+                  <TimeField
                     label={t('events.eventTimeStart', 'Start time') as string}
                     value={formData.event_time_start}
-                    onChange={handleInputChange('event_time_start')}
+                    onChange={(v) => setFormData(prev => ({ ...prev, event_time_start: v }))}
                   />
-                  <Input
-                    type="time"
-                    step={900}
+                  <TimeField
                     label={t('events.eventTimeEnd', 'End time') as string}
                     value={formData.event_time_end}
-                    onChange={handleInputChange('event_time_end')}
+                    onChange={(v) => setFormData(prev => ({ ...prev, event_time_end: v }))}
                   />
                 </div>
               )}
